Day9 completed

+++ AdventOfCode2020/Days/Day1/ValuesMultiplicator.cs
@@ -8,7 +8,7 @@ namespace AdventOfCode2020.Days.Day1
 {
     public class ValuesMultiplicator
     {
-        public static string GetMathOperation(List<int> values)
+        public static string GetMathOperation(List<long> values)
         {
             var result = "";
             foreach (var item in values.Take(values.Count - 1))
